refactor: move creation of div to CreateElement.js
- Move creation of div to CreateElement.js
- Move constants.js and elements.js into common/
