+++ src/common/CreateElement.js
@@ -1,0 +1,3 @@
+export default function createDiv() {
+  return document.createElement('div');
+}
